+++ HISTORY.rst
@@ -10,6 +10,7 @@ History
 - FEAT: allow decorating of default(fallback) handler
 - CHORE: more test coverage
 - CHORE: refactor endpoint urls to settings
+- CHORE: refactor message attributes from SNSMessage to base class
 
 0.1.1 (2019-03-22)
 ++++++++++++++++++